Sorted issues with the creation, updating and deletion of instances

+++ lib.php
@@ -19,26 +19,43 @@
 
 function soundcloud_add_instance($instance)
 {
-  global $DB;
-  
-  
-  return $instance->id;
+    global $DB;
+
+    $instance->timemodified = time();
+    $instance->timecreated  = $instance->timemodified;
+
+    if (!($instance->id = $DB->insert_record('soundcloud', $instance))) {
+        return false;
+    }
+
+    return $instance->id;
 }
 
 function soundcloud_update_instance($instance)
 {
-  global $DB;
-  
-  return $DB->update_record("", $instance);
+    global $DB;
+
+    $instance->timemodified = time();
+    $instance->id = $instance->instance;
+
+    if (!($registration->id = $DB->insert_record('registration', $registration))) {
+        return false;
+    }
+
+    return $DB->update_record('soundcloud', $instance);
 }
 
 function soundcloud_delete_instance($id)
 {
-  global $DB;
-  
-  if (!$instance = $DB->get_record('soundcloud', array('id' => $id))) {
-    return false;
-  }
-  
-  return $DB->delete_records('soundcloud', array('id' => $instance->id));
+    global $DB;
+
+    if (!$instance = $DB->get_record('soundcloud', array('id' => $id))) {
+        return false;
+    }
+
+     if (!$DB->delete_records('soundcloud_posts', array('soundcloud' => $instance->id))) {
+      return false;
+    }
+
+    return $DB->delete_records('soundcloud', array('id' => $instance->id));
 }
